Add files via upload
admittedly slightly hippie-dippy way of fixing this but tbh at this point i just want this bug to be over with.

if self.legistar_start(event) is None, then make the start a passed time that will register the meeting on the Documenters.org website as 'passed' (i.e. a meeting that is already passed and shouldn't be assigned).